requirements.txt
torch
torchvision
torchmetrics[image]
numpy
matplotlib
pyyaml
tqdm
kaggle
wandb
python-dotenv
diffusers
accelerate
huggingface_hub
safetensors
optuna
scikit-learn
pandas